% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_labraduck.R
\name{labraduck_fit}
\alias{labraduck_fit}
\alias{labraduck}
\title{Interface to fit DLM models}
\usage{
labraduck(Y = NULL, upsilon = NULL, Xi = NULL, F, G, W, M0, C0,
  observations, gamma_scale = 0, W_scale = 0, init = NULL,
  pars = c("Eta", "Sigma", "Thetas_filtered", "Thetas_smoothed",
  "M_star", "Eta_DLM"), ...)
}
\arguments{
\item{Y}{D x N matrix of counts (if NULL uses priors only)}

\item{upsilon}{dof for inverse wishart prior (numeric must be > D)
(default: D+3)}

\item{Xi}{(D-1)x(D-1) prior covariance matrix
(default: ALR transform of diag(1)*(upsilon-D)/2 - this is
essentially iid on "base scale" using Aitchison terminology)}

\item{F}{Observation-level transformation matrix (see model)}

\item{G}{System-level transformation matrix (see model)}

\item{W}{System-level noise covariance (of dim Q) over rows of eta transpose (see model)}

\item{M0}{Prior mean over system matrix}

\item{C0}{Prior covariance over system-level noise}

\item{observations}{Vector timepoints indicating offset from first observation (labeled 1)
(ex: 1, 2, 3, 6, 7, 8 where there is a gap in observations)}

\item{init}{(D-1) x Q initialization for Eta for optimization}

\item{pars}{character vector of posterior parameters to return}

\item{...}{}

\item{scale_init}{Scale initialization system (W) and observation level (gamma) noise}

\item{m}{object of class labraduckfit}
}
\value{
an object of class labraduckfit
}
\description{
Labraduck - exchangeable dynamic linear models in stray
}
\details{
the full model is given by:
\deqn{Y_j \sim Multinomial(Pi_j)}
\deqn{Pi_j = Phi^{-1}(Eta_j)}
\deqn{Eta_t^T = F_t^T Theta_t + v_t^T, v_t \sim N_{D-1}(0, gamma_t Sigma)}
\deqn{Theta_t^T = G_t Theta_{t-1} + w_t, w_t \sim N_{Q x D-1}(0, W_t, Sigma)}
\deqn{Sigma \sim InvWish(upsilon, Xi)}
}
