% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labraduckfit_s3.R
\name{labraduckfit}
\alias{labraduckfit}
\title{Create labraduckfit object}
\usage{
labraduckfit(D, N, T, coord_system, iter = NULL, alr_base = NULL,
  ilr_base = NULL, Eta = NULL, Sigma = NULL, Sigma_default = NULL,
  Y = NULL, upsilon = NULL, Xi = NULL, Xi_default = NULL,
  gamma = NULL, F = NULL, G = NULL, W = NULL, M0 = NULL,
  C0 = NULL, observations = NULL, init = NULL)
}
\description{
Create labraduckfit object
}
