% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_dist}
\alias{mat_dist}
\title{Calculate distances between positive-definite covariance matrices}
\usage{
mat_dist(A, B, use_Riemann)
}
\description{
Calculate distances between positive-definite covariance matrices
}
